requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.36.2
antlr4-python3-runtime==4.9.3
anyio==4.4.0
APScheduler==3.10.4
asgiref==3.8.1
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
audioread==3.0.1
av==14.4.0
babel==2.16.0
backports-datetime-fromisoformat==2.0.3
backports.cached-property==1.0.2
bcrypt==4.2.0
beautifulsoup4==4.13.4
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.2
captcha==0.7.1
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
cloudpayments==1.6.3
cmake==3.25.0
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
cryptography==43.0.1
ctranslate2==4.5.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
deepgram-sdk==3.7.5
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
einops==0.8.0
environs==11.0.0
et_xmlfile==2.0.0
exceptiongroup==1.2.1
fal_client==0.5.6
fastapi==0.112.0
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fluent-compiler==0.3
fluent.syntax==0.19.0
fluentogram==1.1.10
fonttools==4.54.1
fpdf==1.7.2
fpdf2==2.8.1
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
google-api-core==2.10.1
google-api-python-client==2.147.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.56.4
greenlet==3.2.1
groq==1.0.0
grpcio==1.67.1
h11==0.14.0
h2==4.2.0
hmmlearn==0.2.8
hpack==4.1.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.1.0
HyperPyYAML==1.2.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.9.2
lightning==2.1.2
lightning-fabric==2.1.2
lightning-utilities==0.11.8
lit==15.0.7
llvmlite==0.43.0
lxml==6.0.0
m3u8==6.0.0
magic-filter==1.0.12
maxapi==0.9.13
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
nats-py==2.9.0
networkx==2.8.8
nltk==3.9.1
numba==0.60.0
numpy==1.23.5
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.20.0
openai==2.20.0
openpyxl==3.1.2
optuna==4.1.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
primePy==1.3
protobuf==3.20.1
psycopg2-binary==2.9.9
pyaes==1.6.1
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.7.4
pydantic_core==2.18.4
pyDeprecate==0.3.2
pydub==0.25.1
pyee==12.1.1
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.2.0
PyPDF2==3.0.1
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytorch-metric-learning==2.8.1
pytubefix==8.2.0
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.4.3
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
semver==3.0.2
sentencepiece==0.2.0
setuptools-rust==1.10.2
shellingham==1.5.4
simplejson==3.19.3
singledispatchmethod==1.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.7
soxr==0.5.0.post1
speechbrain==0.5.16
SQLAlchemy==2.0.31
sshtunnel==0.4.0
starlette==0.37.2
stripe==11.3.0
sympy==1.13.1
tabulate==0.9.0
Telethon==1.36.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.4
transformers==4.39.3
trio==0.30.0
trio-websocket==0.12.2
# triton==2.0.0  # GPU-only, not needed on CPU server
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.30.5
watchdog==2.3.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
yt-dlp==2025.3.31
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0
PyPDF2==3.0.1
